Update README.md
Karina Sonawane • Portfolio is a personal website showcasing Karina’s skills and projects as a Java Full Stack and Web Developer. It features sections for projects, skills, experience, education, and contact. The site is responsive, modern, and easy to navigate, with a dark/light theme toggle and a contact form.